At this point, I think Verlet integration (without velocity) is the only solution to all these physics problems I've been having. I just need to implement it properly.
All I was using velocity for was checking to see which direction the player was going in. On the other hand, knowing the player's previous position would be so useful for doing collision detection. Since the velocity can be easily determined by subtracting the previous position from the position, there's no reason to actually store the velocity.