Moved IsSolid to util.
Layer tiles now default to -1 for each member. That way, we can know for sure to not draw a tile if there is none.

We now draw tile hitboxes in the same way as for the player. Interestingly, the only tile this seems to work for is the one at the top left, despite the fact that there should not be any tile there. After looking into this, it turns out the *only* tile that *is* set as collide-able is the top-left one. How weird!

Fixed NormInt16.

Added assert to ArenaAllocRaw to make sure the size is always more than 0.