Added static to every function, since I am only using them in one translation unit. I believe this saves time for the linker?
I'm honestly not sure what I just spent the past two and a half hours doing. I was so productive for the first hour and a half, and after that I just started going down a rabbit-hole about sub-stepping, even though it was already working fine before. I really need to learn when to stop!