Removed SpriteCellType.
The way it worked before was each sprite cell was either a sprite, a hitbox, a hurtbox, or an origin, determined by the name of the corresponding layer. That is, if the name of the layer was "Hitbox" then the cell would be of type Hitbox, and so on. The only exception was Sprite, which is just the default if any layer name is used besides "Hitbox," "Hurtbox" or "Origin."

The main problem with this approach is that every single time you want to get the hitbox of a sprite, you have to loop through all the cells. This is stupid and unnecessary because each frame can only have one hitbox, one hurtbox and one origin. It would therefore be an error if multiple cells were found that had the same SpriteCellType (besides SpriteCellType_Sprite).

More than that, it also makes it very annoying to loop through the sprite cell array, which is actually the reason why I'm doing this. It would be so nice to just assume that every sprite cell contains a sprite (duh?), but you just couldn't the way it worked before.

Unfortunately, in order to do this, I had to move make a lot of changes, which I know contradicts what I said yesterday. The problem is, the longer you don't hold yourself to any standard when it comes to commits, the harder it gets to make the transition.

Anyways, here it goes:

- Moved Rect to before any other struct definition since it gets used by so many other structs.
- Removed SpriteCellType.
- Removed type from SpriteCell.
- Added hitbox to SpriteFrame.
- Added hurtbox to SpriteFrame. FYI, I still have yet to actually add hurtboxes, since it's not that high priority when nothing can even render right now.
- Added origin to SpriteFrame.
- Added frame_idx parameter to GetSpriteOrigin.
- GetSpriteOrigin no longer loops through the cells of the first frame of the sprite. Instead, it gets the origin from the current frame directly, flipping it if necessary. Speaking of which, why were we only getting the origin from the first frame? That's blatantly incorrect.
- GetSpriteHitbox no longer loops through the cells of the current frame. Instead, it gets the hitbox from the current frame directly, and flips it if necessary. It also adds the origin. Should we do this right after LoadSprites instead?
- LoadSprites now assigns the hitbox to each sprite frame.
- LoadSprites now assigns the origin to each sprite frame.
- Changed the way VulkanCreateImages sets the number of array layers for each sprite in order to match the above changes. Specifically, we no longer have to check each and every sprite cell to see if it is of type SpriteCellType_Sprite.

Dang! I really need to start splitting these up more from now on. Only thing is, only committing some of these changes would have resulted in a lot of things breaking. Maybe I just need to find a good middle-ground?

Next problem to solve: GetEntityHitbox has an assertion failure now.