Added anim and pos back to Entity. Removed inst from Entity. Fixed accidentally returning a pointer to the second entity in GetEntities.
I really need to focus on getting things working, not "perfect." There are so few entities, it doesn't really matter if we have to allocate an array of instances every single frame. Doing it that way greatly simplifies because it means that the gameplay code is completely unaffected by the rendering code.