This isn't a "real" commit. I'm only making it because I'm about to make a really big change.
I've figured out the problem. The way I'm loading sprites is just really stupid. I'm doing it in two passes even though I don't actually have to do that, as long as I just use SDL_TellIO and SDL_SeekIO strategically. Because I'm doing it in two passes, it makes it difficult to know if everything is happening as it should. In particular, it makes it difficult to make sure that the correct amount of sprite cells are getting created.