Moved where static staging buffer size gets incremented.
In general, I think it's better to be clear than it is to be "efficient," especially with something as trivial as this. Who cares if we have to loop through all the cells again? It's easier to understand when the static staging buffer size gets set all in the same place.