Removed pointless spall calls.
I suspect the problem is actually in the way I'm loading the tile layers, since I remember having problems with it before I switched over to Vulkan. I'm trying to understand the JSON code, but for some reason it makes no sense to me. I just thought of switching to a different JSON library but of course that would just be procrastination. I really just need to take a second to try to understand this.